interactive/selfchat fix+base class
