+++ parlai/tasks/light_dialog/worlds.py
@@ -7,6 +7,7 @@
 from parlai.core.worlds import create_task
 from parlai.core.worlds import DialogPartnerWorld, validate
 from parlai.tasks.self_chat.worlds import SelfChatWorld as SelfChatBaseWorld
+from parlai.tasks.interactive.worlds import InteractiveWorld as InteractiveBaseWorld
 from parlai.agents.repeat_label.repeat_label import RepeatLabelAgent
 
 from copy import deepcopy
@@ -15,13 +16,8 @@ import pickle
 import os
 
 
-class InteractiveSimpleWorld(DialogPartnerWorld):
-    def __init__(self, opt, agents, shared=None):
-        super().__init__(opt, agents, shared)
-        print("[ loading personas.. ]")
-        self.load_personas()
-
-    def load_personas(self):
+class InteractiveSimpleWorld(InteractiveBaseWorld):
+    def init_contexts(self):
         # Create Light data so we can assign personas.
         light_opt = self.opt.copy()
         light_opt['task'] = 'light_dialog'
@@ -30,7 +26,7 @@ class InteractiveSimpleWorld(DialogPartnerWorld):
         self.light_world = create_task(light_opt, light_agent)
         self.cnt = 0
 
-    def get_new_personas(self):
+    def get_contexts(self):
         # Find a new episode
         while True:
             self.light_world.parley()
@@ -58,7 +54,7 @@ class InteractiveSimpleWorld(DialogPartnerWorld):
             + '\n'
             + '_self_persona I am a '
             + ' '.join(p['_partner_name'].split(' ')[1:])
-            + '.\n'
+            + '.'
         )
 
         a2_persona = (
@@ -72,45 +68,8 @@ class InteractiveSimpleWorld(DialogPartnerWorld):
             + p['_self_name']
             + '\n'
             + p['_self_persona']
-            + '\n'
         )
         return a1_persona, a2_persona
-
-    def parley(self):
-        """
-        Agent 0 goes first.
-
-        Alternate between the two agents.
-        """
-        if self.cnt == 0:
-            self.p1, self.p2 = self.get_new_personas()
-
-        acts = self.acts
-        agents = self.agents
-        if self.cnt == 0:
-            # add the persona on to the first message to agent 0
-            act = {}
-            act['text'] = self.p1
-            act['episode_done'] = False
-            act['id'] = 'persona'
-            agents[0].observe(validate(act))
-        act = deepcopy(agents[0].act())
-        acts[0] = act
-        if self.cnt == 0:
-            # add the persona on to the first message to agent 1
-            act.force_set('text', self.p2 + act.get('text', 'hi'))
-            agents[1].observe(validate(act))
-        else:
-            agents[1].observe(validate(act))
-        acts[1] = agents[1].act()
-        agents[0].observe(validate(acts[1]))
-        self.update_counters()
-        self.cnt += 1
-
-        if act['episode_done']:
-            print("CHAT DONE ")
-            print("\n... preparing new chat... \n")
-            self.cnt = 0
 
 
 class InteractiveSelfchatWorld(SelfChatBaseWorld):
